🟢 gusto: move the provider onto the v14 module path
The provider was written before the `go.mondoo.com/mql/v13` prefix was
dropped. Every file rebased onto main without conflicting while still
declaring and importing a module path that no longer exists.
