+++ providers/gusto/gen/main.go
@@ -4,8 +4,8 @@
 package main
 
 import (
-	"go.mondoo.com/mql/v13/providers-sdk/v1/plugin/gen"
-	"go.mondoo.com/mql/v13/providers/gusto/config"
+	"go.mondoo.com/mql/providers-sdk/v1/plugin/gen"
+	"go.mondoo.com/mql/providers/gusto/config"
 )
 
 func main() {
