+++ providers/gusto/main.go
@@ -6,8 +6,8 @@ package main
 import (
 	"os"
 
-	"go.mondoo.com/mql/v13/providers-sdk/v1/plugin"
-	"go.mondoo.com/mql/v13/providers/gusto/provider"
+	"go.mondoo.com/mql/providers-sdk/v1/plugin"
+	"go.mondoo.com/mql/providers/gusto/provider"
 )
 
 func main() {
