+++ providers/gusto/resources/gusto.go
@@ -6,9 +6,9 @@ package resources
 import (
 	"context"
 
-	"go.mondoo.com/mql/v13/llx"
-	"go.mondoo.com/mql/v13/providers-sdk/v1/plugin"
-	"go.mondoo.com/mql/v13/providers/gusto/connection"
+	"go.mondoo.com/mql/llx"
+	"go.mondoo.com/mql/providers-sdk/v1/plugin"
+	"go.mondoo.com/mql/providers/gusto/connection"
 )
 
 func (g *mqlGusto) id() (string, error) {
