+++ providers/gusto/resources/gusto.lr
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: BUSL-1.1
 
 option provider = "go.mondoo.com/mql/providers/gusto"
-option go_package = "go.mondoo.com/mql/v13/providers/gusto/resources"
+option go_package = "go.mondoo.com/mql/providers/gusto/resources"
 
 // Gusto
 //
